Lesson 2: Homework tasks 1-14

+++ src/main/java/lv/javaguru/student_miroslavs_fjodorovs/lesson_2/Task_1.java
@@ -15,7 +15,7 @@ public class Task_1 {
        System.out.println(firstNumber + secondNumber);
        System.out.println(firstNumber - secondNumber);
        System.out.println(firstNumber * secondNumber);
-       System.out.println(firstNumber / secondNumber);
+       System.out.println(firstNumber / (double) secondNumber);
